Fixing return
